requirements.txt
streamlit
langchain
langchain-openai
sqlalchemy
pyodbc
python-dotenv
pandas